+++ src/RedisProxyException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace RedisProxy;
+
+use Exception;
+
+class RedisProxyException extends Exception
+{
+}